Match DiskGravity::updateLocalParam, DiskGravity::calcOwnGravityVec, and TRot3f::setRotate
DiskGravity::updateLocalParam uses an inline of TRot3f::setRotate which
is almost (but not quite yet) unified with the original function.
DiskGravity::updateLocalParam also involved a somewhat ugly abuse of
references to obtain the correct codegen. JMath::TSinCosTable has fixed
so that the table is the correct size and the likely intent of the
template paramater is expressed.