Match 99% of `SurfRay` (#1784)
* Match 99% of `FindingLuigiEventScheduler`

 - Match 99% of `FindingLuigiEventScheduler` with only mismatches of string order in `.data` and length of `.rodata`
- Renamed _0 argument of `GalaxyMoveArgument` to `mMoveType`

* Removed extra fields in `FindingLuigiEventScheduler`

* Added future todo about .data mismatch in `FindingLuigiEventScheduler`

* Resolve requested changes in `FindingLuigiEventScheduler`

Formatting should now adhere to project specifications, and I have made fixes to the requested changes in `GalaxyMoveArgument.hpp` and `FindingLuigiEventScheduler.hpp`.
I also reformatted my bracing from "start on next line" to "start on same line" as I see in the rest of the project.
Let me know if there is anything else you need me to do, or if any issues have arisen. :)

* Majority of work matching SurfRay (draft)

Still lots of minor cleanup to do, but just to get the draft out there

* Fix minor mistakes and mark non-matching sections

* Address fixes to draft